# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Sales Platform API
  version: 2.0.0
  description: Authenticated, tenant-isolated CRM, reporting, settings, and AI workflows.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /research/status:
    get:
      operationId: getResearchProviderStatus
      tags: [Research]
      summary: Return sanitized search configuration state
      responses:
        "200": { description: Configuration state returned without secrets }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /research/jobs:
    get:
      operationId: listResearchJobs
      tags: [Research]
      summary: List owned structured research jobs
      responses:
        "200": { description: Research job page }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createResearchJob
      tags: [Research]
      summary: Run a confirmed provider search and evidence-grounded analysis
      responses:
        "201": { description: Structured research job completed }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: refresh_token }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
...
  responses:
...
    Unauthorized: { description: Authentication failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    Conflict: { description: Resource state conflict, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    RateLimited: { description: Request rate limit exceeded, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    ServiceUnavailable: { description: Required dependency unavailable, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }